exports: forcing `_ur` and `_ur_addons` packages as type:module, remove from main package.json
The source files in these directories are typescript modules, so enforcing here is ok.
However, the main package export is mixed, so exclude it there.
This change causes vscode's linter to complain about imports that do not include extensions,
as type:module forces node into requiring extensions to indicate what kind of module is being
imported. A corresponding change is required in tsconfig.json to allowImportingTsExtensions.
For packaging with rollup, however, these extensions have to be removed in its output by a
preprocessing step in the rollup plugin chain (not part of this commit)